refactor: relocate erc20 elated benchmark
 - add test_erc20_operation file
 - move test_sload_erc20_generic, test_sstore_erc20_generic, test_mixed_sload_sstore to same file
