+++ tests/benchmark/stateful/bloatnet/test_single_opcode.py
@@ -38,8 +38,6 @@ from execution_testing import (
 from execution_testing.base_types.base_types import Number
 
 from tests.benchmark.stateful.helpers import (
-    APPROVE_SELECTOR,
-    BALANCEOF_SELECTOR,
     CacheStrategy,
     build_cache_strategy_blocks,
     build_delegated_storage_setup,
@@ -720,230 +718,6 @@ def test_sstore_bloated(
         runtime_code=runtime_code,
         cache_strategy=cache_strategy,
         tx_generator=tx_generator,
-    )
-
-
-@pytest.mark.stub_parametrize(
-    "erc20_stub", "test_sload_empty_erc20_balanceof_"
-)
-def test_sload_erc20_generic(
-    benchmark_test: BenchmarkTestFiller,
-    pre: Alloc,
-    fork: Fork,
-    gas_benchmark_value: int,
-    tx_gas_limit: int,
-    erc20_stub: str,
-) -> None:
-    """Benchmark SLOAD using ERC20 balanceOf on bloatnet."""
-    # Stub Account
-    erc20_address = pre.deploy_contract(
-        code=Bytecode(),
-        stub=erc20_stub,
-    )
-    threshold = 100000
-
-    # MEM[0] = function selector
-    # MEM[32] = starting address offset
-    setup = Op.MSTORE(
-        0,
-        BALANCEOF_SELECTOR,
-        # gas accounting
-        old_memory_size=0,
-        new_memory_size=32,
-    ) + Op.MSTORE(
-        32,
-        Op.SLOAD(0),  # Address Offset
-        # gas accounting
-        old_memory_size=32,
-        new_memory_size=64,
-    )
-
-    call_balance_of = Op.POP(
-        Op.CALL(
-            address=erc20_address,
-            args_offset=32 - 4,
-            args_size=32 + 4,
-        )
-    )
-
-    loop = While(
-        body=call_balance_of + Op.MSTORE(32, Op.ADD(Op.MLOAD(32), 1)),
-        condition=Op.GT(Op.GAS, threshold),
-    )
-
-    teardown = Op.SSTORE(0, Op.MLOAD(32))
-
-    # Contract Deployment
-    code = setup + loop + teardown
-    attack_contract_address = pre.deploy_contract(code=code)
-
-    intrinsic_gas = fork.transaction_intrinsic_cost_calculator()()
-
-    # Transaction Loops
-    txs = []
-    gas_remaining = gas_benchmark_value
-
-    sender = pre.fund_eoa()
-
-    while gas_remaining > intrinsic_gas:
-        gas_available = min(gas_remaining, tx_gas_limit)
-
-        if gas_available < intrinsic_gas:
-            break
-
-        with TestPhaseManager.execution():
-            txs.append(
-                Transaction(
-                    gas_limit=gas_available,
-                    to=attack_contract_address,
-                    sender=sender,
-                )
-            )
-
-        gas_remaining -= gas_available
-
-    blocks = [Block(txs=txs)]
-    benchmark_test(
-        pre=pre,
-        blocks=blocks,
-        skip_gas_used_validation=True,
-        expected_receipt_status=True,
-    )
-
-
-# SLOAD BENCHMARK ARCHITECTURE:
-#
-#   [Pre-deployed ERC20 Contract] ──── Storage slots for balances
-#           │
-#           │  balanceOf(address) → SLOAD(keccak256(address || slot))
-#           │
-#   [Attack Contract] ──CALL──► ERC20.balanceOf(random_address)
-#           │
-#           └─► Loop(i=0 to N):
-#                 1. Generate random address from counter
-#                 2. CALL balanceOf(random_address) → forces cold SLOAD
-#                 3. Most addresses have zero balance → empty storage slots
-#
-# WHY IT STRESSES CLIENTS:
-#   - Each balanceOf() call forces a cold SLOAD on a likely-empty slot
-#   - Storage slot = keccak256(address || balances_slot)
-#   - Random addresses ensure maximum cache misses
-#   - Tests client's sparse storage handling efficiency
-
-
-# SSTORE BENCHMARK ARCHITECTURE:
-#
-#   [Pre-deployed ERC20 Contract] ──── Storage slots for allowances
-#           │
-#           │  approve(spender, amount)
-#           │    → SSTORE(keccak256(spender || slot), amount)
-#           │
-#   [Attack Contract]
-#       ──CALL──► ERC20.approve(counter_as_spender, counter_as_amount)
-#           │
-#           └─► Loop(i=0 to N):
-#                 1. Use counter as both spender address and amount
-#                 2. CALL approve(counter, counter) → forces cold SSTORE
-#                 3. Writes to new allowance slots in sparse storage
-#
-# WHY IT STRESSES CLIENTS:
-#   - Each approve() call forces an SSTORE to a new storage slot
-#   - Storage slot = keccak256(
-#       msg.sender || keccak256(spender || allowances_slot)
-#     )
-#   - Sequential counter ensures unique storage locations
-#   - Tests client's ability to handle many storage writes
-#   - Simulates real-world contract state accumulation over time
-
-
-@pytest.mark.stub_parametrize("erc20_stub", "test_sstore_erc20_approve_")
-def test_sstore_erc20_generic(
-    benchmark_test: BenchmarkTestFiller,
-    pre: Alloc,
-    fork: Fork,
-    gas_benchmark_value: int,
-    tx_gas_limit: int,
-    erc20_stub: str,
-) -> None:
-    """Benchmark SSTORE using ERC20 approve."""
-    sender = pre.fund_eoa()
-
-    threshold = 100_000
-
-    # Stub Account
-    erc20_address = pre.deploy_contract(
-        code=Bytecode(),
-        stub=erc20_stub,
-    )
-
-    # MEM[0] = function selector
-    # MEM[32] = starting address offset
-    setup = Op.MSTORE(
-        0,
-        APPROVE_SELECTOR,
-    ) + Op.MSTORE(
-        32,
-        Op.SLOAD(0),  # Address Offset
-    )
-
-    call_approve = Op.MSTORE(
-        64,
-        Op.ADD(1, Op.MLOAD(32)),
-    ) + Op.POP(
-        Op.CALL(
-            address=erc20_address,
-            args_offset=28,
-            args_size=68,
-        )
-    )
-
-    loop = While(
-        body=call_approve + Op.MSTORE(32, Op.ADD(Op.MLOAD(32), 1)),
-        condition=Op.GT(Op.GAS, threshold),
-    )
-
-    teardown = Op.SSTORE(0, Op.MLOAD(32))
-
-    # Contract Deployment
-    code = setup + loop + teardown
-    attack_contract_address = pre.deploy_contract(code=code)
-
-    intrinsic_gas = fork.transaction_intrinsic_cost_calculator()()
-
-    # Transaction Loops
-    gas_remaining = gas_benchmark_value
-
-    # Collect tx params first, then build Transaction objects
-    # so that nonces are allocated contiguously per block.
-    tx_gas: list[int] = []
-    while gas_remaining > intrinsic_gas:
-        gas_available = min(gas_remaining, tx_gas_limit)
-
-        if gas_available < intrinsic_gas:
-            break
-
-        tx_gas.append(gas_available)
-
-        gas_remaining -= gas_available
-
-    txs = []
-    with TestPhaseManager.execution():
-        for gas_available in tx_gas:
-            txs.append(
-                Transaction(
-                    gas_limit=gas_available,
-                    to=attack_contract_address,
-                    sender=sender,
-                )
-            )
-
-    blocks = [Block(txs=txs)]
-
-    benchmark_test(
-        pre=pre,
-        blocks=blocks,
-        skip_gas_used_validation=True,
-        expected_receipt_status=True,
     )
